+++ ceammc/ext/src/flow/flow_speedlim.h
@@ -3,6 +3,7 @@
 
 #include "ceammc_clock.h"
 #include "ceammc_object.h"
+#include "ceammc_proxy.h"
 
 using namespace ceammc;
 
@@ -10,6 +11,7 @@ class FlowSpeedLimit : public BaseObject {
     IntProperty* period_;
     bool is_closed_;
     ClockMemberFunction<FlowSpeedLimit> clock_;
+    InletProxy<FlowSpeedLimit> inlet_;
 
 public:
     FlowSpeedLimit(const PdArgs& a);
@@ -20,7 +22,9 @@ public:
     void onAny(t_symbol* s, const AtomListView& l) override;
     void onData(const Atom& data) override;
 
-    void m_reset(t_symbol*, const AtomListView&);
+    void proxy_float(t_float f) { period_->setValue(f); }
+    void proxy_reset() { proxy_reset({}); }
+    void proxy_reset(const AtomListView&);
 
     void accept();
     void clock_handler();